refactor: go/0kn/{internal -> pkg}/utils
