+++ go/0kn/cmd/xtrellis/xtrellis.go
@@ -6,7 +6,7 @@ import (
 
 	arg "github.com/alexflint/go-arg"
 
-	"github.com/31333337/bmrng/go/0kn/internal/utils"
+	"github.com/31333337/bmrng/go/0kn/pkg/utils"
 	"github.com/31333337/bmrng/go/trellis/client"
 	"github.com/31333337/bmrng/go/trellis/config"
 	"github.com/31333337/bmrng/go/trellis/errors"
